Move user registration validation to userService
